pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "beat-montage-engine"
version = "0.1.0"
description = "Beat-synchronized video montage generator using aesthetic-indexer database"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [{ name = "delora" }]
dependencies = [
    "typer>=0.12.0",
    "pydantic>=2.6.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    
    "librosa>=0.10.2",
    "numpy>=1.26.0",
    "scipy>=1.11.0",
    "soundfile>=0.12.1",
    
    "SQLAlchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "pgvector>=0.3.6",
    
    "pyyaml>=6.0",
]

[project.scripts]
beat-montage = "src.pipeline.run:app"

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

